Use "${var:?}" to ensure this never expands to /* . Added config version check

+++ clamav-unofficial-sigs.sh
@@ -874,7 +874,7 @@ function remove_script () {
           #rather keep the configs
           #rm -f -- "$default_config" && echo "     Removed file: $default_config"
           #rm -f -- "$0" && echo "     Removed file: $0"
-          xshok_is_subdir "$work_dir" && rm -rf -- "$work_dir" && echo "     Removed script working directories: $work_dir"
+          xshok_is_subdir "$work_dir" && rm -rf -- "${work_dir:?}" && echo "     Removed script working directories: $work_dir"
 
           echo "  The clamav-unofficial-sigs script and all of its associated files, third-party"
           echo "  databases, and work directories have been successfully removed from the system."
@@ -1143,6 +1143,22 @@ function check_new_version () {
     fi
   fi
 }
+
+#function to check for a new version
+function check_new_config_version () {
+  if [ "$wget_bin" != "" ] ; then
+    latest_config_version="$($wget_bin https://raw.githubusercontent.com/extremeshok/clamav-unofficial-sigs/master/config/master.conf -O - 2> /dev/null | $grep_bin  "config_version=" | cut -d\" -f2)"
+  else
+    latest_config_version="$($curl_bin https://raw.githubusercontent.com/extremeshok/clamav-unofficial-sigs/master/config/master.conf 2> /dev/null | $grep_bin  "config_version=" | cut -d\" -f2)"
+  fi
+  if [ "$latest_config_version" ] ; then
+    if [ ! "$latest_config_version" == "$config_version" ] ; then
+      xshok_pretty_echo_and_log "New configversion : v$latest_config_version @ https://github.com/extremeshok/clamav-unofficial-sigs" "-"
+    fi
+  fi
+}
+
+
 
 #function for help and usage
 ##usage: 
@@ -3101,6 +3117,8 @@ clamscan_reload_dbs
 xshok_pretty_echo_and_log "Issue tracker : https://github.com/extremeshok/clamav-unofficial-sigs/issues" "-"
 
 check_new_version
+
+check_new_config_version
 
 xshok_cleanup
